diffusion_model/requirements.txt
absl-py
matplotlib==3.9.2
numpy==2.1.2
Pillow==11.0.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
wandb==0.18.5
opencv-python==4.11